+++ runtime/source/core/timer.cpp
@@ -3,7 +3,7 @@
 
 namespace z1 {
 
-    double Timer::fixed_update_delta = 0.016;
+    float Timer::fixed_update_delta = 0.016;
 
     Timer::Timer() {
         m_prev_timepoint = m_start_timepoint = std::chrono::steady_clock::now();
@@ -11,7 +11,7 @@ namespace z1 {
 
     void Timer::update() {
         auto now = std::chrono::steady_clock::now();
-        auto delta_time_count = std::chrono::duration<double, std::micro>(now - m_prev_timepoint).count();
+        auto delta_time_count = std::chrono::duration<float, std::micro>(now - m_prev_timepoint).count();
         m_delta_time = 0.000001 * delta_time_count;
         m_prev_timepoint = now;
 
@@ -25,12 +25,12 @@ namespace z1 {
         }
     }
 
-    double Timer::get_delta_time() const {
+    float Timer::get_delta_time() const {
         return m_delta_time;
     }
 
-    double Timer::get_total_time() const {
-        auto total_time_count = std::chrono::duration<double, std::micro>(m_prev_timepoint - m_start_timepoint).count();
+    float Timer::get_total_time() const {
+        auto total_time_count = std::chrono::duration<float, std::micro>(m_prev_timepoint - m_start_timepoint).count();
         return 0.000001 * total_time_count;
     }
